requirements.txt
pyTelegramBotAPI
googletrans==4.0.0-rc1
python-dotenv
requests
Flask
gunicorn
waitress